+++ src/Krp.ReverseProxy/KubernetesForwarder/EndpointExplorer/EndpointExplorerBackgroundService.cs
@@ -1,4 +1,4 @@
-using Krp.KubernetesForwarder.Dns;
+using Krp.KubernetesForwarder.Routing;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
