+++ src/Krp.ReverseProxy/KubernetesForwarder/HttpForwarder/HttpForwarderBackgroundService.cs
@@ -5,7 +5,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Krp.KubernetesForwarder;
+namespace Krp.KubernetesForwarder.HttpForwarder;
 
 public class HttpForwarderBackgroundService : BackgroundService
 {
