+++ src/Krp.ReverseProxy/KubernetesForwarder/Routing/DnsWindowsHostsHandler.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Krp.KubernetesForwarder.Dns;
+namespace Krp.KubernetesForwarder.Routing;
 
 public interface IDnsHandler
 {
